feat(table): add style func for custom cell rendering
This brings back the styleFunc introduced in [1].

I didn't run into the alignment issues that were the reason this was
removed[2]. The lipgloss package also has table support, which the bubbles
maintainers plan to eventually switch to, but IMO since this is a widget it
doesn't belong in lipgloss, which is concerned with style definitions.

I also plan to add additional features to the table, and prefer doing
this here.

[1]: https://github.com/charmbracelet/bubbles/pull/504

[2]: https://github.com/charmbracelet/bubbles/pull/586